Fixes issue #1
Some images didn't get converted properly. This was easy to
spot in smaller image where the last part seemed to be missing.
Each of the hex values generated stands for a byte = 8 bits =
8 pixels. But sometimes the size of the image caused the last
hex value to contain less than 8 bits (because there weren't
enough pixels left). As the hex values represent 8 pixels being drawn
right to left, the missing bits would cause the pixels to move out
of the canvas.
The solution is to always pad the last hex to a length of 8.
.