Fixes problem with small images
This should finally fix the problem with small images getting distorted.
Problem appears to have been that with some sizes (basically
everything that wasn't divisible by 8) the last part of a row would be
merged with the first part of the last row, creating an offset that
increased with every next row.

This update pads every last byte of a row with zeros to fill out to 8
bits.

Closing #21