Support ESP+u8g2 with optional bitswap
Somewhere (SPI Interface?) there is a MSB/LSB swap when using the converter with ESP32 and the U8g2 library 
which distorts the output.
I checked with 1bit per pixel output but added an optional bitswap checkbox, a function to do optinal swap and 
added it to all output functions.
The lvgl online converter has got the same problem. Maybe there is is something wrong with the u8g...
Since endian is a traditional an generic issue I thought it's probably a nice feature inside the converter 
which at least works fine for me.